+++ docutils/docutils/writers/xetex/__init__.py
@@ -24,6 +24,7 @@ __docformat__ = 'reStructuredText'
 
 from docutils import frontend
 from docutils.writers import latex2e
+from docutils.writers.latex2e import PreambleCmds
 
 
 class Writer(latex2e.Writer):
@@ -145,3 +146,18 @@ class XeLaTeXTranslator(latex2e.LaTeXTranslator):
         else:
             self.requirements['_inputenc'] = (r'\XeTeXinputencoding %s '
                                               % self.latex_encoding)
+
+    def to_latex_length(self, length_str: str) -> str:
+        """Convert "measure" `length_str` to LaTeX length specification.
+
+        XeTeX does not know the length unit px.
+        Use ``\\pdfpxdimen``, the macro holding the value of 1 px in pdfTeX.
+        This way, configuring works the same for pdftex and xetex.
+        """
+        length_str = super().to_latex_length(length_str)
+        if length_str.endswith('px'):
+            if not self.fallback_stylesheet:
+                self.fallbacks['_providelength'] = PreambleCmds.providelength
+            self.fallbacks['px'] = '\n\\DUprovidelength{\\pdfpxdimen}{1bp}\n'
+            return length_str.replace('px', '\\pdfpxdimen')
+        return length_str